IW-1636 | rename Notifications & Search components
