+++ source/components/GlobalNavigation/components/Notifications/MarkAsReadIcon.js
